fixed erro
there was an erro when while in brazil was 1am EUA was -1
because the diferencaHorario was negative, so brazil 1am + (-2 EUA) = -1. so i just added a condition to check these cases, and the hour will return the value as it should. Also added the day -= 1, so you will return to yesterday.